Revert "omap4xxx: Fix aliasing violations, optimize"
This reverts commit 84909ea5840dfb45c7bd6dcc5c86f99d671617db.

+++ camera/AppCallbackNotifier.cpp
@@ -1061,13 +1061,13 @@ void AppCallbackNotifier::notifyFrame()
                                 void *y_uv[2];
                                 mapper.lock((buffer_handle_t)vBuf, CAMHAL_GRALLOC_USAGE, bounds, y_uv);
 
-                                structConvImage input =  {(int)frame->mWidth,
-                                                          (int)frame->mHeight,
+                                structConvImage input =  {frame->mWidth,
+                                                          frame->mHeight,
                                                           4096,
                                                           IC_FORMAT_YCbCr420_lp,
                                                           (mmByte *)frame->mYuv[0],
                                                           (mmByte *)frame->mYuv[1],
-                                                          (int)frame->mOffset};
+                                                          frame->mOffset};
 
                                 structConvImage output = {mVideoWidth,
                                                           mVideoHeight,
